BROWSE3 REFACTOR - Massive file migration
