+++ weave-js/src/components/PagePanelComponents/Home/Browse2/Browse2RootObjectVersionUsers.tsx
@@ -4,7 +4,7 @@ import {useHistory, useParams} from 'react-router-dom';
 
 import {callsTableFilter, callsTableNode, callsTableOpCounts} from './callTree';
 import {Browse2RootObjectVersionItemParams} from './CommonLib';
-import {useWeaveflowRouteContext} from './context';
+import {useWeaveflowRouteContext} from '../Browse3/context';
 import {opDisplayName} from './dataModel';
 import {LinkTable} from './LinkTable';
