+++ weave-js/src/components/PagePanelComponents/Home/Browse3/pages/CallPage.tsx
@@ -1,6 +1,6 @@
 import React, {useMemo} from 'react';
 
-import {Browse2TraceComponent} from '../Browse2TracePage';
+import {Browse2TraceComponent} from '../../Browse2/Browse2TracePage';
 import {SimplePageLayout} from './common/SimplePageLayout';
 import {useWeaveflowORMContext} from './interface/wf/context';
