+++ weave-js/src/components/PagePanelComponents/Home/Browse3/pages/ObjectsPage.tsx
